+++ src/db/migrations/0004_abnormal_bloodstrike.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `comics` ADD `deleted_at` integer;